backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.3.2
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.3
django-appconf==1.0.4
django-compressor==2.4
django-cors-headers==3.5.0
django-heroku==0.3.1
django-picklefield==3.0.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
isort==5.6.4
joblib==0.17.0
lazy-object-proxy==1.4.3
logilab-common==1.8.0
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.4
pep8==1.5.7
Pillow==8.0.1
psycopg2==2.8.6
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
rcssmin==1.0.6
requests==1.1.0
requests-toolbelt==0.9.1
rjsmin==1.1.0
scikit-learn==0.23.2
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.1
whitenoise==5.2.0
wrapt==1.12.1
xlwt==1.3.0
yahoo==0.1
yql-finance==0.1.0
ystockquote==0.2.5
